+++ src/run/ui_main.py
@@ -1,6 +1,7 @@
 import os
 import sys
 
+
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from PyQt5.QtWidgets import QApplication
@@ -8,6 +9,7 @@ from loguru import logger
 from ui_event.base_event import exit_save_data
 from log.log_record import log_record, check_version
 from ui_event.pyqt_main_ui import UIMainWindows
+from utils.sytem_monitor import sys_mon
 
 
 @logger.catch
@@ -23,6 +25,7 @@ def ui_paint():
 
     w = UIMainWindows()
     w.show()
+    sys_mon()
     app.exec_()